Split the widget post menu tests into a separate file
